fixed exam thing

+++ new-exam.sh
@@ -5,11 +5,6 @@ if [ "$#" -ne 2 ]; then
 	exit 1
 fi
 
-if [ -d "$1" ]; then
-	echo "Directory $1 already exists. Exiting..."
-	exit 1
-fi
-
 month=`date +%m`
 year=`date +%Y`
 
@@ -17,9 +12,15 @@ if [ "$year" -lt 6 ]; then
 	$[ year-- ]
 fi
 
-mkdir "$1"
-cp "template.tex" "$1/CSC-$1-$2-$year-Review.tex"
-cp "template-ANSWERS.tex" "$1/CSC-$1-$2-$year-Review-ANSWERS.tex"
-mkdir "$1/questions"
-cp "template-makefile" "$1/Makefile"
+if [ ! -d "$1" ]; then
+	mkdir "$1"
+	cp "template-makefile" "$1/Makefile"
+fi
+if [ ! -f "$1/CSC-$1-$2-$year-Review.tex" ]; then
+	cp "template.tex" "$1/CSC-$1-$2-$year-Review.tex"
+	cp "template-ANSWERS.tex" "$1/CSC-$1-$2-$year-Review-ANSWERS.tex"
+fi
+if [ ! -d "$1/questions" ]; then
+	mkdir "$1/questions"
+fi
 echo "\\input{CSC-$1-$2-$year-Review.tex}" >> "$1/CSC-$1-$2-$year-Review-ANSWERS.tex"
